requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
SQLAlchemy==2.0.49
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pandas==3.0.0
numpy==2.4.2
